requirements.txt
openai
anthropic
google-genai
requests
pandas
ipython # For display() and HTML()
matplotlib
seaborn
numpy
rich
pathlib
